Create Blank DB Table Row When SendEmail
When sending email to student, a blank row will be created in the forms table with forms id= student id and responded = 0. When the Student Submits the form, the table will be updated with the input and responded = 1.

Next step: Change the flagging status and the color of the flag